Refactor to include simplecov-latest in simplecov.go
